Support running apps as standalone microservices.
Setting the EG_SERVICE_AS_MICRO environment variable when starting a
service launches it as a microservice.

Typically, a service runs as a threaded team of workers with a single
controlling thread.  In microservice mode, only a single, standalone
worker is started.  The intention here is to better support integration
with orchestration services like k8s.

Related, service methods are now stored in a OnceLock for easier global
access.

Also simplified the names of some service modules.

Signed-off-by: Bill Erickson <berickxx@gmail.com>