Bus timeout no longer supports block-forever
Calling recv() on the Redis bus with an indefinite timeout leads to
scenarios where signals cannot be handled, and should generally be
avoided.  Related, the Timer can no longer return a negative amount for
time remaining, which was leading to unexpected block-forever scenarios
on recv().  Instead, if a client needs to wait essentially forever, then
wrap the recv() in a loop.

Signed-off-by: Bill Erickson <berickxx@gmail.com>