SIP2 mediator gets aliveness-account setting/signals
Adds an aliveness-account setting to the config file.  When present,
this account will not be allowed to login to the mediator if the
mediator is in non-ready mode.

Currently, ready mode is on by default, disabled with a SIGUSR1 to the
mediator, and re-enabled with a SIGUSR2 to the mediator.

Signed-off-by: Bill Erickson <berickxx@gmail.com>