Print a redacted form of the DSN to STDERR when we fail to connect to the db
When you can't connect to the db, it's helpful to know exactly which params we
were using to attempt the connection.  This info was already being logged at the
debug level, but I thought it might be nice to provide the info right away on
error (except for the password, for security reasons).

Signed-off-by: Jane Sandberg <js7389@princeton.edu>