[marctk] Add optional convenience methods for Marc21 bibliographic records
This commit adds two new methods behind a feature flag named
marc21_bibliographic (with the thought that future contributors
might want to add convenience methods for, say, authority records
or UNIMARC bibliographic records).

The `cargo test` invocation in CI is updated to run tests for
all features.

We also enable the nightly doc_auto_cfg feature when docs.rs
generates the marctk documentation, which provides nice
feature flag badges in the documentation for any functionality
that is behind a flag.

Signed-off-by: Jane Sandberg <js7389@princeton.edu>